requirements.txt
Django==3.2.10
django-cors-headers==3.11.0
django-filter==2.4.0
djangorestframework==3.13.1
